+++ workflows_definition/RELEASES.md
@@ -219,3 +219,11 @@ Based on:
 - Speakeasy CLI 1.451.5 (2.471.2) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.11.2] workflows_definition
+
+## 2024-12-10 00:32:11
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.453.8 (2.474.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.11.3] workflows_definition